Implement CI selfcheck script (fixed). (#99)

+++ CIScripts/Test/Tests/Protocol/SingleHostWithUtils.Tests.ps1
@@ -1,5 +1,5 @@
 Param (
-    [Parameter(Mandatory=$true)] [string] $TestenvConfFile,
+    [Parameter(Mandatory=$false)] [string] $TestenvConfFile,
     [Parameter(Mandatory=$false)] [string] $LogDir = "pesterLogs"
 )
 
@@ -14,14 +14,6 @@ Param (
 . $PSScriptRoot\..\..\PesterHelpers\PesterHelpers.ps1
 . $PSScriptRoot\..\..\Utils\CommonTestCode.ps1 # Get-RemoteNetAdapterInformation
 . $PSScriptRoot\..\..\Utils\DockerImageBuild.ps1 
-
-$Sessions = New-RemoteSessions -VMs (Read-TestbedsConfig -Path $TestenvConfFile)
-$Session = $Sessions[0]
-
-$OpenStackConfig = Read-OpenStackConfig -Path $TestenvConfFile
-$ControllerConfig = Read-ControllerConfig -Path $TestenvConfFile
-$SystemConfig = Read-SystemConfig -Path $TestenvConfFile
-$IisTcpTestDockerImage = "iis-tcptest"
 
 Describe "Single compute node protocol tests with utils" {
 
@@ -161,8 +153,19 @@ Describe "Single compute node protocol tests with utils" {
         }
     }
 
-
     BeforeAll {
+        $Sessions = New-RemoteSessions -VMs (Read-TestbedsConfig -Path $TestenvConfFile)
+        $Session = $Sessions[0]
+
+        $OpenStackConfig = Read-OpenStackConfig -Path $TestenvConfFile
+        $ControllerConfig = Read-ControllerConfig -Path $TestenvConfFile
+        [Diagnostics.CodeAnalysis.SuppressMessageAttribute(
+            "PSUseDeclaredVarsMoreThanAssignments", "",
+            Justification="Analyzer doesn't understand relation of Pester blocks"
+        )]
+        $SystemConfig = Read-SystemConfig -Path $TestenvConfFile
+        $IisTcpTestDockerImage = "iis-tcptest"
+
         Initialize-DockerImage -Session $Session -DockerImageName $IisTcpTestDockerImage
 
         Install-DockerDriver -Session $Session
@@ -178,6 +181,8 @@ Describe "Single compute node protocol tests with utils" {
     }
 
     AfterAll {
+        if (-not (Get-Variable Sessions -ErrorAction SilentlyContinue)) { return }
+
         Write-Host "Removing iis-tcptest image from testbed"
         Invoke-Command -Session $Session {
             docker image rm $Using:IisTcpTestDockerImage -f 2>$null
@@ -186,7 +191,7 @@ Describe "Single compute node protocol tests with utils" {
         Uninstall-DockerDriver -Session $Session
         Uninstall-Extension -Session $Session
         Uninstall-Utils -Session $Session
+
+        Remove-PSSession $Sessions
     }
 }
-
-Remove-PSSession $Sessions
